Wold_Virtual/mi-proyecto/requirements.txt
Flask==2.0.1
Flask-Cors==3.9.0
Flask-Sockets==0.2.1
Flask-MySQLdb==1.0.0
web3==5.24.0
eth-account==0.5.6
PyJWT==2.1.0
requests==2.25.1
```